pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "queryhub"
version = "0.1.0"
description = "Automation platform for scheduled query-based email reports."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
  { name = "QueryHub Maintainers" }
]
keywords = ["automation", "reports", "scheduling"]
classifiers = [
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = [
  "PyYAML>=6.0",
  "pydantic>=2.12",
  "Jinja2>=3.1",
  "typer>=0.20",
  "aiosmtplib>=5.0",
  "aiohttp>=3.13",
  "SQLAlchemy[asyncio]>=2.0",
  "asyncpg>=0.30",
  "azure-kusto-data>=5.0",
  "azure-identity>=1.25",
  "boto3>=1.40",
  "google-cloud-bigquery>=3.38",
  "google-auth>=2.43",
  "plotly>=6.4",
  "numpy>=2.3",
  "pandas>=2.3",
  "kaleido>=1.2",
  "python-dateutil>=2.9"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "ruff>=0.5",
  "mypy>=1.7",
  "pytest-asyncio>=0.23",
  "pytest-cov>=4.1",
  "types-PyYAML>=6.0.12.20240808",
  "boto3-stubs[essential]>=1.34",
  "aiosqlite>=0.20",
  "bandit[toml]>=1.8",
  "safety>=3.7"
]

[project.urls]
Homepage = "https://github.com/isasnovich/QueryHub"
Documentation = "https://github.com/isasnovich/QueryHub#readme"
Source = "https://github.com/isasnovich/QueryHub"
Changelog = "https://github.com/isasnovich/QueryHub/blob/main/CHANGELOG.md"

[project.scripts]
queryhub = "queryhub.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/queryhub"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
testpaths = ["tests"]
markers = [
  "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = ["B101"]  # Skip assert_used check (common in tests)
